# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "the server is developed by Go",
        "title": "ToDoList API",
        "termsOfService": "https://github.com/go-programming-tour-book",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "uncertainly search in all the task by key information",
                "parameters": [
                    {
                        "description": "rush",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SearchTaskService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseTask"
                        }
                    },
                    "500": {
                        "description": "status\":500,\"data\":{},\"Msg\":{},\"Error\":\"error\"}",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/task": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update a task",
                "parameters": [
                    {
                        "description": "rush",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateTaskService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseTask"
                        }
                    },
                    "500": {
                        "description": "status\":500,\"data\":{},\"Msg\":{},\"Error\":\"error\"}",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "create a task",
                "parameters": [
                    {
                        "description": "title",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateTaskService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseTask"
                        }
                    },
                    "500": {
                        "description": "status\":500,\"data\":{},\"Msg\":{},\"Error\":\"error\"}",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/task/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "show the detail of one task",
                "parameters": [
                    {
                        "description": "rush",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ShowTaskService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseTask"
                        }
                    },
                    "500": {
                        "description": "status\":500,\"data\":{},\"Msg\":{},\"Error\":\"error\"}",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "uncertainly search in all the task by key information",
                "parameters": [
                    {
                        "description": "rush",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SearchTaskService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseTask"
                        }
                    },
                    "500": {
                        "description": "status\":500,\"data\":{},\"Msg\":{},\"Error\":\"error\"}",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get the task list",
                "parameters": [
                    {
                        "description": "rush",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ListTaskService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseTask"
                        }
                    },
                    "500": {
                        "description": "status\":500,\"data\":{},\"Msg\":{},\"Error\":\"error\"}",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "user_name, password",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"登陆成功\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseUser"
                        }
                    },
                    "500": {
                        "description": "{\"status\":500,\"data\":{},\"Msg\":{},\"Error\":\"error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseUser"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户名, 密码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UserService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseUser"
                        }
                    },
                    "500": {
                        "description": "{\"status\":500,\"data\":{},\"Msg\":{},\"Error\":\"error\"}",
                        "schema": {
                            "$ref": "#/definitions/serializer.ResponseUser"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "serializer.ResponseTask": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/serializer.Task"
                },
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "serializer.ResponseUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/serializer.User"
                },
                "error": {
                    "type": "string",
                    "example": ""
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "serializer.Task": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string",
                    "example": "红烧肉，炒白菜"
                },
                "create_at": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "integer"
                },
                "id": {
                    "description": "任务ID",
                    "type": "integer",
                    "example": 1
                },
                "start_at": {
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0未完成，1已完成）",
                    "type": "integer",
                    "example": 0
                },
                "title": {
                    "description": "题目",
                    "type": "string",
                    "example": "吃饭"
                },
                "view": {
                    "description": "浏览量",
                    "type": "integer",
                    "example": 32
                }
            }
        },
        "serializer.User": {
            "type": "object",
            "properties": {
                "create_at": {
                    "description": "创建",
                    "type": "integer"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "description": "用户状态",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string",
                    "example": "dengzhao"
                }
            }
        },
        "service.CreateTaskService": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "status": {
                    "description": "0是未做 1是已做",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "service.ListTaskService": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "service.SearchTaskService": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "string"
                },
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "service.ShowTaskService": {
            "type": "object"
        },
        "service.UpdateTaskService": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "status": {
                    "description": "0是未做 1是已做",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "service.UserService": {
            "type": "object",
            "required": [
                "password",
                "user_name"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 16,
                    "minLength": 5
                },
                "user_name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        }
    }
}